Refactor createNewUser function to include role-based access control and improve code organization
Related to #1509
