--- src/app/utils/firebase/createNewUser.ts
@@ -1,59 +1,0 @@
-import { db } from "./firebaseConfig";
-import { doc, setDoc } from "firebase/firestore";
-import { AffiliateInfo, UserData } from "../../types";
-import { toast } from "react-toastify";
-
-export async function createNewUser(
-  walletAddress: string,
-  userInfo: AffiliateInfo,
-): Promise<string> {
-  const userDocRef = doc(db, "users", walletAddress);
-
-  const newUser: UserData = {
-    username: userInfo.username,
-    email: userInfo.email,
-    role: userInfo.role,
-    createdAt: new Date(),
-    updatedAt: new Date(),
-    // Set access control based on user role
-    allowed: userInfo.role === "ProjectOwner" ? false : true,
-  };
-
-  // Add projectUrl only if the role is ProjectOwner
-  if (userInfo.role === "ProjectOwner" && userInfo.projectUrl) {
-    newUser.projectUrl = userInfo.projectUrl;
-  }
-
-  // Add joinedProjectIds only if the role is Affiliate
-  if (userInfo.role === "Affiliate") {
-    newUser.joinedProjectIds = [];
-  }
-
-  // Add X Auth Token and X Account Info if they exist
-  if (userInfo.xAuthToken) {
-    newUser.xAuthToken = userInfo.xAuthToken; // Save the X auth token data
-  }
-
-  if (userInfo.xAccountInfo) {
-    newUser.xAccountInfo = userInfo.xAccountInfo; // Save the X account information
-  }
-
-  // Add Google Auth Token and YouTube Account Info if they exist
-  if (userInfo.googleAuthToken) {
-    newUser.googleAuthToken = userInfo.googleAuthToken; // Save the Google auth token data
-  }
-
-  if (userInfo.youtubeAccountInfo) {
-    newUser.youtubeAccountInfo = userInfo.youtubeAccountInfo; // Save the YouTube account information
-  }
-
-  try {
-    await setDoc(userDocRef, newUser);
-    toast.success("Your account has been created successfully!");
-    return walletAddress;
-  } catch (error) {
-    console.error("Failed to create new user and join project: ", error);
-    toast.error("Failed to create new user and join project.");
-    throw new Error("Failed to create new user and join project");
-  }
-}
